Replace the ResponseType model with a field
Instead of using a ResponseType model and a ManyToMany relationship to Client, use a custom field that saves the selected response_types in a JSONField.
This simplifies the model structure and doesn't require keeping the db in-sync with code, since the response_types are a static set of choices anyway.