WIP fresh migrations
Basic migrations working. Trickery needed because of circular dependencies in migrations (hence the first migation only defines Client and the secondd the rest. A app extending the models needs to depend on the first migration, so the replaced models can have a relation to Client (need to edit migration after creation).

old migrations are deleted. But they need to be restored and somehow have the new ones mark them as replaced (not sure if that works well with multiple migrations).